requirements.txt
selenium
undetected-chromedriver
discord-webhook
python-dotenv
flask
APScheduler
Flask-BasicAuth
gunicorn
pypdf